imgrate parsing stuff to parsetasks.cpp
